Refactor and first suite of tests
