+++ src/logger.js
@@ -1,0 +1,10 @@
+const flags = {
+	logging: 1,
+}
+
+function vlog(level, message) {
+	if (flags.logging >= level)
+		console.log(message);
+}
+
+module.exports = { vlog, flags };
